Complete state/memory filesystem implementation via 9P
Expands 9P filesystem with four additional memory access files:
- state/memory/wram (36KB binary, 8 banks)
- state/memory/oam (256B binary, sprite data)
- state/memory/highram (256B binary, I/O registers)
- state/memory/state (text format, banking state with validation)

Key design decisions:
- Generic binaryMemoryFS implementation eliminates ~700 lines of duplication across wram/oam/highram
- Text format for memory state enables partial updates and clear error messages
- Command.Offset changed from int64 to int to match Go slice indexing
- Helper methods (GetWRAM, GetOAM, GetHighRAM, GetMemoryState) follow existing GetVRAM pattern with no internal locking
- CommandChan buffer increased to 32 to support current + future state files (APU, PPU, cartridge)

Testing:
- 25 unit tests (100% pass rate)
- All tests pass with race detector (-race flag)
- Tests cover read/write, validation, edge cases, concurrent access

Thread safety: Reads use RWMutex for consistency. Writes validated synchronously then queued via command channel for processing at frame boundaries (~16ms).

Documentation updated with comprehensive examples for save/restore, binary operations, and text-based state management.

+++ pkg/ninep/vramfs.go
@@ -63,7 +63,7 @@ func (f *vramFile) Write(p []byte) (n int, err error) {
 	// Queue write command for frame boundary processing
 	cmd := gb.Command{
 		Name:   "vram-write",
-		Offset: int64(f.readPos),
+		Offset: f.readPos,
 		Data:   make([]byte, len(p)),
 	}
 	copy(cmd.Data, p)
@@ -106,7 +106,7 @@ func (f *vramFile) WriteAt(p []byte, offset int64) (n int, err error) {
 
 	cmd := gb.Command{
 		Name:   "vram-write",
-		Offset: offset,
+		Offset: int(offset),
 		Data:   make([]byte, len(p)),
 	}
 	copy(cmd.Data, p)
